Minor optimization that will be more noticeable on larger files.

Test bench: split_into_lines.v
```v
import benchmark

lines := '
one
two
three
four
five
six
seven
eight
nine
ten'

mut bench := benchmark.start()
for _ in 0..100000 {
        lines.split_into_lines()
}
bench.measure('split loop')
```
Several runs:
```
jalon@Ripper:~/testing$ v run split_into_lines.v
 SPENT   144.427 ms in split loop
jalon@Ripper:~/testing$ ~/git/vv/v run split_into_lines.v
 SPENT   118.418 ms in split loop
jalon@Ripper:~/testing$ v run split_into_lines.v
 SPENT   142.780 ms in split loop
jalon@Ripper:~/testing$ ~/git/vv/v run split_into_lines.v
 SPENT   119.177 ms in split loop
jalon@Ripper:~/testing$ v run split_into_lines.v
 SPENT   145.559 ms in split loop
jalon@Ripper:~/testing$ ~/git/vv/v run split_into_lines.v
 SPENT   118.264 ms in split loop
jalon@Ripper:~/testing$
```

BONUS:  `unsafe` no longer required.



<!--

Please title your PR as follows: `time: fix foo bar`.
Always start with the thing you are fixing, then describe the fix.
Don't use past tense (e.g. "fixed foo bar").

Explain what your PR does and why.

If you are adding a new function, please document it and add tests:

```
// foo does foo and bar
fn foo() {

// file_test.v
fn test_foo() {
    assert foo() == ...
    ...
}
```

If you are fixing a bug, please add a test that covers it.

Before submitting a PR, please run `v test-all` .
See also `TESTS.md`.

I try to process PRs as soon as possible. They should be handled within 24 hours.

Applying labels to PRs is not needed.

Thanks a lot for your contribution!

-->